refactor(user):rntity > api.domain 경로 수정
